requirements.txt
mss>=9.0.0
Pillow>=10.0.0
keyboard>=0.13.5
pywin32>=306
pystray>=0.19.0
psutil>=5.9.0
win10toast>=0.9
PyQt6>=6.6.0
dxcam>=0.0.5
